разбиения на модули
